requirements.txt

Flask
Flask-Cors
google-cloud-firestore
gunicorn
google-auth
firebase-admin

python-dotenv